requirements.txt
Pillow==10.3.0
streamlit==1.35.0
textblob==0.18.0.post0
transformers==4.41.2
vaderSentiment==3.3.2
vaderSentiment==3.3.2
